requirements.txt
fastapi>=0.104.1
uvicorn[standard]>=0.23.2
sqlmodel>=0.0.14
alembic>=1.12.1
asyncpg>=0.29.0
psycopg2-binary>=2.9.9
aiogram>=3.1.1
python-dotenv>=1.0.0
httpx>=0.25.2
pydantic-settings>=2.0.3
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.6
python-jose[cryptography]>=3.3.0
cryptography>=41.0.0
slowapi>=0.1.9
pytest>=7.4.3
pytest-asyncio>=0.21.1
aiosqlite>=0.20.0
